Move logout endpoint to session resource
