+++ src/back-end/lib/resources/session.ts
@@ -1,8 +1,11 @@
+import axios from 'axios';
+import { KEYCLOAK_CLIENT_ID, KEYCLOAK_CLIENT_SECRET, KEYCLOAK_REALM, KEYCLOAK_URL } from 'back-end/config';
 import * as crud from 'back-end/lib/crud';
-import { Connection, deleteSession } from 'back-end/lib/db';
+import { Connection, deleteSession, readOneSession,  } from 'back-end/lib/db';
 import * as permissions from 'back-end/lib/permissions';
 import { basicResponse, makeJsonResponseBody } from 'back-end/lib/server';
-import { SupportedRequestBodies, SupportedResponseBodies } from 'back-end/lib/types';
+import { ServerHttpMethod, SupportedRequestBodies, SupportedResponseBodies } from 'back-end/lib/types';
+import qs from 'querystring';
 import { Session } from 'shared/lib/resources/session';
 
 type Resource = crud.Resource<SupportedRequestBodies, SupportedResponseBodies, null, null, Session, Connection>;
@@ -32,7 +35,31 @@ const resource: Resource = {
         if (!permissions.deleteSession(request.session, request.params.id)) {
           return respond(401, [permissions.ERROR_MESSAGE]);
         }
-        await deleteSession(connection, request.session.id);
+
+        const session = await readOneSession(connection, request.session.id);
+        if (!session.accessToken) {
+          return respond(200, null);
+        }
+
+        const formData = qs.stringify({
+          client_id: KEYCLOAK_CLIENT_ID,
+          client_secret: KEYCLOAK_CLIENT_SECRET,
+          scope: 'openid',
+          refresh_token: session.accessToken
+        });
+
+        await axios({
+          method: ServerHttpMethod.Post,
+          url: `${KEYCLOAK_URL}/auth/realms/${KEYCLOAK_REALM}/protocol/openid-connect/logout`,
+          data: formData,
+          headers: {
+            'Content-Type': 'application/x-www-form-urlencoded'
+          }
+        });
+
+        // Delete the current session
+        await deleteSession(connection, session.id);
+
         return respond(200, null);
       }
     };
